Change song text size algorithm
The old algorithm was faulty in some cases, as it limited the number of lines to 2, which caused songs with large titles to become too small.

The new algorithm limits the height of the title text to 40% of the album art size, thus not limiting the number of lines of text, but the height of the text as a whole, making text size bigger for long song titles.